Changed instruction size to 64 bits. Added load instructions

+++ src/LibEphedra/LibEphedra/Easm/Instructions/Jmp.cs
@@ -1,10 +1,5 @@
 ﻿using LibEphedra.VM;
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
 namespace LibEphedra.Easm.Instructions
 {
     public class Jmp : Instruction
@@ -15,7 +10,7 @@ namespace LibEphedra.Easm.Instructions
 
         public override void Execute(CPU cpu)
         {
-            var ip = cpu.GetRegister((byte)SpecialRegister.IP);
+            var ip = cpu.GetRegister(SpecialRegister.IP);
             ip.Value = Immediate;
         }
     }
